utils/ai.py
import openai
from typing import Dict, Any, List, Optional
import logging
import json
from datetime import datetime
import asyncio
from concurrent.futures import ThreadPoolExecutor

from config.settings import get_settings

# Initialize settings
settings = get_settings()
logger = logging.getLogger(__name__)

# Initialize OpenAI client
client = openai.OpenAI(api_key=settings.OPENAI_API_KEY)

# Thread pool for running synchronous OpenAI operations asynchronously
executor = ThreadPoolExecutor()

async def generate_lesson_content(
    subject: str,
    topic: str,
    difficulty: str,
    duration_minutes: int,
    additional_instructions: Optional[str] = None
) -> Dict[str, Any]:
    """
    Generate a lesson using OpenAI
    
    Args:
        subject: Subject area (e.g., "Mathematics", "Physics")
        topic: Specific topic (e.g., "Linear Algebra", "Quantum Mechanics")
        difficulty: Difficulty level (e.g., "beginner", "intermediate", "advanced")
        duration_minutes: Expected duration in minutes
        additional_instructions: Any additional instructions for the AI
        
    Returns:
        Dict: Generated lesson content including sections, exercises, etc.
    """
    try:
        # Create system prompt
        system_prompt = f"""
        You are an expert educational content creator specialized in {subject}. 
        Your task is to create a comprehensive, engaging, and educational lesson on {topic}.
        
        The lesson should be appropriate for {difficulty} level students and designed to take approximately {duration_minutes} minutes to complete.
        
        Provide a well-structured lesson with the following components:
        1. An engaging title
        2. A brief overview/summary
        3. 3-7 content sections (depending on lesson length)
        4. 2-5 relevant exercises or quiz questions with answers
        5. A list of additional resources for further learning
        6. Relevant tags for categorization
        
        Format your response as a valid JSON object with the following structure:
        {
            "title": "Lesson Title",
            "summary": "Brief overview of the lesson",
            "content": [
                {
                    "title": "Section Title",
                    "content": "Section content with explanations, examples, etc.",
                    "order": 1,
                    "type": "text"
                }
                // Additional sections...
            ],
            "exercises": [
                {
                    "question": "Exercise question",
                    "options": ["Option A", "Option B", "Option C", "Option D"],
                    "correct_answer": "Correct answer",
                    "explanation": "Explanation of the answer",
                    "difficulty": "medium"
                }
                // Additional exercises...
            ],
            "resources": [
                {
                    "title": "Resource Title",
                    "url": "URL if applicable",
                    "type": "link",
                    "description": "Brief description of the resource"
                }
                // Additional resources...
            ],
            "tags": ["tag1", "tag2", "tag3"]
        }
        """
        
        # Add additional instructions if provided
        if additional_instructions:
            system_prompt += f"\n\nAdditional instructions: {additional_instructions}"
            
        # Run in thread pool to avoid blocking
        def _generate_content():
            response = client.chat.completions.create(
                model=settings.OPENAI_MODEL,
                messages=[
                    {"role": "system", "content": system_prompt},
                    {"role": "user", "content": f"Create a {difficulty} level lesson on {topic} in {subject} that takes about {duration_minutes} minutes to complete."}
                ],
                temperature=0.7,
                max_tokens=4000,
                response_format={"type": "json_object"}
            )
            
            # Parse response content
            content = response.choices[0].message.content
            return json.loads(content)
            
        return await asyncio.get_event_loop().run_in_executor(executor, _generate_content)
        
    except Exception as e:
        logger.error(f"Error generating lesson content: {str(e)}")
        raise

async def generate_answer(
    question: str,
    context: Optional[str] = None,
    lesson_content: Optional[List[Dict[str, Any]]] = None
) -> Dict[str, Any]:
    """
    Generate an answer to a question using OpenAI
    
    Args:
        question: The question to answer
        context: Optional additional context
        lesson_content: Optional lesson content for reference
        
    Returns:
        Dict: Generated answer with references
    """
    try:
        # Create system prompt
        system_prompt = """
        You are an AI tutor assistant that provides helpful, accurate, and educational responses to student questions.
        Your answers should be:
        1. Clear and concise
        2. Accurate and factual
        3. Educational and instructive
        4. Encouraging further learning
        
        If you use specific sources or reference materials in your answer, include them in your response.
        
        Format your response as a valid JSON object with the following structure:
        {
            "answer": "Your comprehensive answer to the question",
            "references": [
                {
                    "title": "Reference Title",
                    "source": "Source name/type",
                    "url": "URL if applicable"
                }
                // Additional references if used...
            ]
        }
        """
        
        # Create user prompt
        user_prompt = question
        
        # Add context if available
        if context:
            user_prompt += f"\n\nAdditional context: {context}"
            
        # Add lesson content if available
        if lesson_content:
            lesson_text = "\n\nRelevant lesson content:\n"
            for section in lesson_content:
                if isinstance(section, dict):
                    section_title = section.get("title", "Untitled Section")
                    section_content = section.get("content", "")
                    lesson_text += f"--- {section_title} ---\n{section_content}\n\n"
            user_prompt += lesson_text
            
        # Run in thread pool to avoid blocking
        def _generate_answer():
            response = client.chat.completions.create(
                model=settings.OPENAI_MODEL,
                messages=[
                    {"role": "system", "content": system_prompt},
                    {"role": "user", "content": user_prompt}
                ],
                temperature=0.5,
                max_tokens=2000,
                response_format={"type": "json_object"}
            )
            
            # Parse response content
            content = response.choices[0].message.content
            return json.loads(content)
            
        return await asyncio.get_event_loop().run_in_executor(executor, _generate_answer)
        
    except Exception as e:
        logger.error(f"Error generating answer: {str(e)}")
        raise

async def generate_text_to_speech(text: str) -> bytes:
    """
    Generate speech from text using Google Text-to-Speech
    
    Args:
        text: The text to convert to speech
        
    Returns:
        bytes: Audio data in MP3 format
    """
    try:
        # Limit text length to avoid exceeding API limits
        if len(text) > 5000:
            text = text[:5000] + "..."
            
        # This is a placeholder for the actual TTS implementation
        # In a real implementation, we would use Google Cloud TTS here
        
        # Example implementation using Google Cloud TTS:
        """
        from google.cloud import texttospeech
        
        def _generate_tts():
            # Instantiates a client
            client = texttospeech.TextToSpeechClient()
            
            # Set the text input to be synthesized
            synthesis_input = texttospeech.SynthesisInput(text=text)
            
            # Build the voice request
            voice = texttospeech.VoiceSelectionParams(
                language_code=settings.TTS_LANGUAGE_CODE,
                name=settings.TTS_VOICE_NAME,
                ssml_gender=texttospeech.SsmlVoiceGender.NEUTRAL
            )
            
            # Select the type of audio file
            audio_config = texttospeech.AudioConfig(
                audio_encoding=texttospeech.AudioEncoding.MP3
            )
            
            # Perform the text-to-speech request
            response = client.synthesize_speech(
                input=synthesis_input, voice=voice, audio_config=audio_config
            )
            
            # Return the audio content
            return response.audio_content
        
        return await asyncio.get_event_loop().run_in_executor(executor, _generate_tts)
        """
        
        # For now, return a placeholder
        logger.warning("TTS functionality is not fully implemented")
        return b''
        
    except Exception as e:
        logger.error(f"Error generating text-to-speech: {str(e)}")
        raise

